+++ CHANGELOG.md
@@ -12,6 +12,7 @@ BUG FIXES:
 * mdb: ensure that `network_id` is not empty if provided
 * mdb: fix host creating with empty `subnet_id` in `yandex_mdb_postgresql_cluster`
 * mdb: suppress master nodes config changes if no master node used in `yandex_mdb_elasticsearch_cluster`
+* apigateway: fixed spec updating in `yandex_api_gateway`
 
 ENHANCEMENTS:
 * add `deletion_protection` attribute to `yandex_mdb_clickhouse_cluster`, `yandex_mdb_elasticsearch_cluster`, `yandex_mdb_kafka_cluster`, `yandex_mdb_mongodb_cluster`, `yandex_mdb_mysql_cluster`, `yandex_mdb_postgresql_cluster`,  `yandex_mdb_redis_cluster`, `yandex_mdb_sqlserver_cluster` resources and data sources